pyproject.toml
[project]
name = "sbomify-action"
version = "0.12"
description = "Generate, augment, enrich, and manage SBOMs in your CI/CD pipeline"
authors = [{ name = "sbomify", email = "hello@sbomify.com" }]
requires-python = ">=3.11"
readme = "README.md"
license = "Apache-2.0"
keywords = ["sbom", "cyclonedx", "spdx", "supply-chain", "security", "bom", "software-composition-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "requests>=2.32.3,<3",
    "cyclonedx-bom>=7.2.1,<8",
    "cyclonedx-python-lib>=11.5.0,<12",
    "sentry-sdk>=2.21.0,<3",
    "spdx-tools>=0.8.3",
    "rich>=14.2.0",
    "zstandard>=0.25.0",
    "packageurl-python>=0.17.6",
    "conan>=2.0,<3",
    "questionary>=2.0.1,<3",
]

[project.urls]
Homepage = "https://sbomify.com"
Repository = "https://github.com/sbomify/github-action"
Documentation = "https://github.com/sbomify/github-action#readme"
"Bug Tracker" = "https://github.com/sbomify/github-action/issues"
Changelog = "https://github.com/sbomify/github-action/releases"

[project.scripts]
sbomify-action = "sbomify_action.cli.main:main"
sbomify-license-db = "sbomify_action._enrichment.license_db_generator:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0,<9",
    "pytest-cov>=4.1.0,<5",
    "pytest-mock>=3.12.0,<4",
    "ruff>=0.12.0,<0.13",
    "pre-commit>=4.2.0,<5",
    "jsonschema>=4.25.1",
    "twine>=6.2.0",
]

[tool.hatch.build.targets.sdist]
include = [
    "sbomify_action/**/*.py",
    "sbomify_action/**/*.json",
]
core-metadata-version = "2.3"

[tool.hatch.build.targets.wheel]
packages = ["sbomify_action"]
core-metadata-version = "2.3"

[tool.hatch.build]
include = [
    "sbomify_action/**/*.py",
    "sbomify_action/**/*.json",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=sbomify_action --cov-report=term-missing"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
filterwarnings = [
    "default::DeprecationWarning",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]  # Ignore line length

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

